+++ src/io/ft_printf/putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: gd-hallu <gd-hallu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/15 12:31:32 by gd-hallu          #+#    #+#             */
-/*   Updated: 2026/01/06 10:56:36 by gd-hallu         ###   ########.fr       */
+/*   Updated: 2026/06/30 12:28:58 by gd-hallu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 int	putstr_fd(char *s, int fd)
 {
     if (!s)
-        return (write(fd, "(null)", 6));
-    return (write(fd, s, ft_strlen((const char *) s)));
+        return (int)(write(fd, "(null)", 6));
+    return (int)(write(fd, s, ft_strlen((const char *) s)));
 }
